♻️ extract mockWorker
