+++ packages/rum/test/index.ts
@@ -1,1 +1,2 @@
+export * from './mockWorker'
 export * from './utils'
